requirements.txt
langgraph>=0.2.0
openai>=1.35.0
arxiv>=2.1.0
python-pptx>=0.6.23
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
redis>=5.0.0
langgraph-checkpoint-redis>=0.1.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
httpx>=0.27.0
streamlit>=1.36.0
python-multipart>=0.0.9
requests>=2.32.0
pymupdf>=1.24.0
pdfplumber>=0.10.0
pdf2image>=1.17.0
Pillow>=10.0.0